forgotten files

+++ models/databse.py
@@ -93,7 +93,7 @@ class MovieRating(Base):
     id = Column(Integer, primary_key=True)
     movie_pick_id = Column(Integer, ForeignKey("movie_picks.id"), nullable=False)
     rater_user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    rating = Column(Float, nullable=False)
+    rating = Column(Integer, nullable=False)
     review_text = Column(Text)
     rated_at = Column(DateTime, default=func.now())
